refactor: collect Rank implementations in one file
