--- src/main/kotlin/blackjack/domain/Ace.kt
@@ -1,5 +1,0 @@
-package blackjack.domain
-
-object Ace : Rank {
-    override val possibleValues: Set<Int> = setOf(1, 11)
-}
